Renaming {About,Home} pages components; for lazy loading the module must be default exported, instead we getting errors
