--- src/components/pages/About/About.js
@@ -1,5 +1,0 @@
-import React from 'react'
-
-const About = () => <h1>About Page</h1>
-
-export default About
