+++ src/components/pages/About/index.js
@@ -1,1 +1,5 @@
-export { default as About } from './About'
+import React from 'react'
+
+const About = () => <h1>About Page</h1>
+
+export default About
