--- src/components/pages/Home/Home.js
@@ -1,15 +1,0 @@
-import React from 'react'
-
-import { Form, TasksList } from '#root/components/tasks'
-
-import { Title } from './styled'
-
-const Home = () => (
-  <>
-    <Title>Home Page</Title>
-    <Form />
-    <TasksList />
-  </>
-)
-
-export default Home
