+++ src/components/pages/Home/index.js
@@ -1,1 +1,15 @@
-export { default as Home } from './Home'
+import React from 'react'
+
+import { Form, TasksList } from '#root/components/tasks'
+
+import { Title } from './styled'
+
+const Home = () => (
+  <>
+    <Title>Home Page</Title>
+    <Form />
+    <TasksList />
+  </>
+)
+
+export default Home
