--- src/components/pages/index.js
@@ -1,2 +1,0 @@
-export { Home } from './Home'
-export { About } from './About'
